Rename the dir name and add the unittest dir
