+++ examples/langchian_llms_mock/langchian_llms_mock.py
@@ -4,7 +4,7 @@ from gptcache.adapter.langchain_llms import LangChainLLMs
 from langchain.llms import OpenAI
 from gptcache.core import cache, Config
 from gptcache.cache.factory import get_ss_data_manager
-from gptcache.ranker.simple import pair_evaluation
+from gptcache.similarity_evaluation.simple import pair_evaluation
 from gptcache.processor.post import nop as postnop
 from gptcache.processor.pre import nop as prenop
 import numpy as np
@@ -18,7 +18,7 @@ def mock_embeddings(data, **kwargs):
 
 
 def run():
-    sqlite_file = "sqlite.db"
+    sqlite_file = "gptcache.db"
     faiss_file = "faiss.index"
     has_data = os.path.isfile(sqlite_file) and os.path.isfile(faiss_file)
     data_manager = get_ss_data_manager("sqlite", "faiss",
