+++ examples/openai_request/openai_request.py
@@ -3,8 +3,8 @@ import time
 
 from gptcache.cache.factory import get_data_manager, get_ss_data_manager
 from gptcache.core import cache, Cache, Config
-from gptcache.encoder import Towhee
-from gptcache.ranker.simple import SearchDistanceEvaluation
+from gptcache.embedding import Towhee
+from gptcache.similarity_evaluation.simple import SearchDistanceEvaluation
 from gptcache.adapter import openai
