+++ gptcache/cache/vector_data/milvus.py
@@ -1,4 +1,4 @@
-from gptcache.util import import_pymilvus
+from gptcache.utils import import_pymilvus
 import_pymilvus()
 
 from uuid import uuid4
