+++ gptcache/embedding/openai.py
