+++ gptcache/embedding/string.py
