+++ gptcache/embedding/towhee.py
@@ -1,4 +1,4 @@
-from gptcache.util import import_towhee
+from gptcache.utils import import_towhee
 import_towhee()
 
 import numpy as np
