+++ gptcache/utils/error.py
